Move Card component into RelatedWorks
